+++ src/components/widget/widget.jsx
@@ -3,22 +3,29 @@ import classNames from 'classnames';
 import './widget.scss';
 
 function renderHeader(header) {
+  let headerBody = header;
+
   if (typeof header === 'string') {
-    return <h4 className="widget__heading">{ header }</h4>;
+    headerBody = <h4 className="widget__heading">{ header }</h4>;
   }
 
-  return header;
+  return (
+    <div className="widget__header">
+      { headerBody }
+    </div>
+  );
 }
 
 export default function Widget({ header, children, fullWidth, className, ...rest }) {
   const classes = classNames('widget', className);
   const bodyClasses = classNames('widget__body', {
     'widget__body--full-width': fullWidth,
+    'widget__body--no-header': !header,
   });
 
   return (
     <div { ...rest } className={ classes }>
-      <div className="widget__header">{ renderHeader(header) }</div>
+      { header ? renderHeader(header) : <span /> }
       <div className={ bodyClasses }>
         { children }
       </div>