+++ src/index.js
@@ -1,17 +1,35 @@
 import Alert from './components/alert/alert';
 import Badge from './components/badge/badge';
 import Button from './components/button/button';
+import DatePicker from './components/date-picker';
 import Dropdown from './components/dropdown/dropdown';
 import Graph from './components/graph/graph';
 import Icon from './components/icon/icon';
 import Input from './components/input/input';
 import Logo from './components/logo/logo';
 import Pane from './components/pane/pane';
+import RatioBar from './components/ratio-bar';
 import Select from './components/select/select';
-import variables from './utilities/variables';
-import Value from './components/value/value';
-import DatePicker from './components/date-picker';
-import Widget from './components/widget/widget';
 import SparkGraph from './components/spark-graph';
+import Value from './components/value/value';
+import variables from './utilities/variables';
+import Widget from './components/widget/widget';
 
-export { Alert, Badge, Button, DatePicker, Dropdown, Graph, Icon, Input, Logo, Pane, Select, SparkGraph, Value, Widget, variables };
+export {
+  Alert,
+  Badge,
+  Button,
+  DatePicker,
+  Dropdown,
+  Graph,
+  Icon,
+  Input,
+  Logo,
+  Pane,
+  RatioBar,
+  Select,
+  SparkGraph,
+  Value,
+  variables,
+  Widget,
+};